sync

+++ lowstepper_daisy_firmware/lowstepper/AverageBuffer.cpp
@@ -22,7 +22,7 @@ void AverageBuffer<T>::addValue(T value) {
 template <class T>
 T AverageBuffer<T>::getAverageValue(void) {
   T sum = 0;
-  for(int i = 0; i < this->bufferSize; i++) {
+  for(size_t i = 0; i < this->bufferSize; i++) {
     sum += values[i];
   }
